Split native shell runtime state support
